Wrote out some conditionals for requests that would use them in actionRoute

+++ actionRoute/index.js
@@ -15,7 +15,7 @@ router.get("/", async (req, res) => {
 router.get("/:id", async (req, res) => {
   const { id } = req.params;
   try {
-    const action = db.get(id);
+    const action = await db.get(id);
     if (action) {
       res.status(200).json(action);
     } else {
@@ -29,7 +29,12 @@ router.get("/:id", async (req, res) => {
 router.post("/", async (req, res) => {
   const action = req.body;
   try {
-    db.insert(action);
+    const newAction = await db.insert(action);
+    if (newAction) {
+      res.status(200).json(newAction);
+    } else {
+      res.status(500).json({ message: "error" });
+    }
   } catch (error) {
     res.status(500).json({ message: "error" });
   }
@@ -38,7 +43,13 @@ router.post("/", async (req, res) => {
 router.delete("/:id", async (req, res) => {
   const { id } = req.params;
   try {
-    db.remove(id);
+    const action = await db.get(id);
+    if (action) {
+      const deleted = await db.remove(id);
+      res.status(200).json(action);
+    } else {
+      res.status(404).json({ message: "error" });
+    }
   } catch (error) {
     res.status(500).json({ message: "error" });
   }
@@ -48,7 +59,12 @@ router.put("/:id", async (req, res) => {
   const { id } = req.params;
   const action = req.body;
   try {
-    db.update(id, action);
+    const edited = await db.update(id, action);
+    if (edited) {
+      res.status(200).json(edited);
+    } else {
+      res.status(404).json({ message: "error" });
+    }
   } catch (error) {
     res.status(500).json({ message: "error" });
   }
